feat(utils): Time difference
