+++ myportfy-frontend/src/components/PostComponent/index.jsx
@@ -3,6 +3,9 @@ import { useState } from "react";
 // assets
 import { BsThreeDotsVertical } from 'react-icons/bs';
 
+// utils
+import { timeDifference } from "../../utils/time-difference";
+
 // styles
 import * as Styled from "./PostStyles";
 
@@ -14,37 +17,6 @@ export const Post = ({ props }) => {
   const toggleBtn = () => {
     setShowMore((prevState) => !prevState);
   };
-
-  function dateFormat(d) {
-    const dateNow = new Date();
-    const postDate = new Date(d);
-    const postFormattedDate = new Intl.DateTimeFormat("pt-Br", {
-      dateStyle: "short",
-    }).format(Date.parse(d));
-
-    //get the difference in minutes
-    if (dateNow.getTime() - postDate.getTime() <= 3600000) {
-      var diffMinutes = Math.ceil(Math.abs(dateNow - postDate) / (1000 * 60));
-      return `Há ${diffMinutes} minuto(s)`;
-    }
-
-    //get the difference in hours
-    else if (dateNow.getTime() - postDate.getTime() < 86400000) {
-      var diffHours = Math.ceil(
-        Math.abs(dateNow - postDate) / (1000 * 3600) - 1
-      );
-      return `Há ${diffHours} hora(s)`;
-    }
-
-    //get the difference in days
-    else if ((dateNow - postDate) / (1000 * 3600 * 24) - 1 <= 31) {
-      var diffDays = Math.ceil(
-        Math.abs(dateNow - postDate) / (1000 * 3600 * 24) - 1
-      );
-      return `Há ${diffDays} dia(s)`;
-    }
-    return postFormattedDate;
-  }
 
   return (
     <Styled.Container>
@@ -86,7 +58,7 @@ export const Post = ({ props }) => {
         </Styled.BoostDiv>
 
         <Styled.Texts fontSmall={true}>
-          {dateFormat(props.createdAt)}
+          {timeDifference(props.createdAt)}
         </Styled.Texts>
       </Styled.BottomDiv>
     </Styled.Container>
